% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{fun_annot}
\alias{fun_annot}
\title{Add annotation}
\usage{
fun_annot(p, x = NULL, y = NULL, text = NULL)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{x, y}{coordinate.}

\item{text}{annotation.}
}
\description{
Add annotation.
}
\examples{
funplot() \%>\%
  fun_add("x^2") \%>\%
  fun_annot(x = 1, text = "x = 1") \%>\%
  fun_annot(y = 2, text = "y = 2")

}
