% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fun.R
\name{fun_deriv}
\alias{fun_deriv}
\title{Add derivative}
\usage{
fun_deriv(p, fun, mouse = FALSE)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{fun}{see details.}

\item{mouse}{update tip on mouse move.}
}
\description{
Add a derivative to the previously added function.
}
\details{
If \code{mouse} is set to true then tangent line is computed whenever the mouse is moved inside the canvas
(let \code{x_{0}} be the mouse's abscissa then the tangent line to the poin
\eqn{(x_{0},f(x_{1}))} is computed whenever the position of the mouse changes).
}
\examples{
funplot() \%>\%
  fun_add("x^2") \%>\%
  fun_deriv("2 * x", mouse = TRUE)

funplot() \%>\%
  fun_add("x * x") \%>\%
  fun_deriv("2 * x", mouse = TRUE) \%>\%
  fun_add("x * x * x") \%>\%
  fun_deriv("3 * x * x", mouse = TRUE)

}
