% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fun_grid}
\alias{fun_grid}
\title{Add grid}
\usage{
fun_grid(p, grid = TRUE)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{grid}{Set to \code{TRUE} add grid (default).}
}
\description{
Add plot grid.
}
\examples{
funplot() \%>\%
  fun_add("(x ^ 2) - 3") \%>\%
  fun_grid() \%>\%
  fun_title("Show grid")

}
