% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fun.R
\name{fun_math}
\alias{fun_math}
\title{Add math.js}
\usage{
fun_math(p, fun, ...)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{fun}{Function to plot.}

\item{...}{Any other parameter, see \code{\link{fun_add}}.}
}
\description{
Add a Add function to be evaluated by math.js.
}
\details{
\code{funplot} uses interval-arithmetic math by default, unfortunately some functions are not
implemented yet because of the underlying complexity, for this reason you can always evaluate a function with
\href{mathjs}{http://mathjs.org/}.
}
\examples{
funplot() \%>\%
 fun_math("gamma(x)")

funplot() \%>\%
 fun_math("atan2(x, x ^ 2) / pi", samples = 4000) \%>\%
 fun_y(domain = list(-1, 1))

}
