% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fun.R
\name{fun_secants}
\alias{fun_secants}
\title{Add secants}
\usage{
fun_secants(p, x0, x1 = NULL, mouse = FALSE)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{x0, x1}{see details.}

\item{mouse}{update secant on mouse move.}
}
\description{
Add secants to the previously added function.
}
\details{
If a data object has a secants, then each object will be used to compute secant lines between two
points belonging to the function, additionally if \code{updateOnMouseMove} is a property set to true in the object then
\eqn{(x_{0},f(x_{1}))} will be used as an anchored point and \eqn{(x_{0},f(x_{1}))} will be computed dynamically
based on the mouse abscissa.

Values for \code{secants} list:
\itemize{
  \item{\code{x0} }{the abscissa of the first point}
  \item{\code{x1} }{(optional if \code{updateOnMouseMove} is set) the abscissa of the second point}
  \item{\code{mouse} }{ (optional) if set to \code{TRUE} \code{x1} will be computed dynamically
  based on the current position of the mouse}.
}
}
\examples{
funplot() \%>\%
  fun_add("x^2") \%>\%
  fun_secants(x0 = 5, mouse = TRUE) \%>\%
  fun_secants(x0 = 2, x1 = 3) \%>\%
  fun_y(domain = list(-200, 400)) \%>\%
  fun_x(domain = list(-300, 300))

}
