% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fun_size}
\alias{fun_size}
\title{Change size}
\usage{
fun_size(p, height = NULL, width = NULL)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{height, width}{Dimentions of plot.}
}
\description{
Change plot size.
}
\examples{
funplot() \%>\%
  fun_add("tan(x)") \%>\%
  fun_size(200, 400) \%>\%
  fun_title("Plot size")

}
