% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fun_tip}
\alias{fun_tip}
\title{Customise tooltip}
\usage{
fun_tip(p, x = FALSE, y = FALSE, renderer = NULL)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{x, y}{Set to \code{TRUE} to show x and y lines.}

\item{renderer}{Custom rendering function for the text shown in the tip.}
}
\description{
Customise the tooltip.
}
\details{
\code{x = TRUE} shows a dashed line parallel to \eqn{y=0} on the tip position,
\code{y = TRUE} shows a dashed line parallel to \eqn{x=0} on the tip position
}
\examples{
funplot() \%>\%
  fun_add("(x ^ 2) - 3") \%>\%
  fun_tip(TRUE, TRUE)

}
