% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fun_vector}
\alias{fun_vector}
\title{Add a vector}
\usage{
fun_vector(p, vector, offset = NULL, ...)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{vector}{A vector of length 2.}

\item{offset}{A vector of displacement from the origin.}

\item{...}{Any other parameter.}
}
\description{
Add a vector to plot.
}
\examples{
funplot() \%>\%
  fun_vector(c(2, 1), c(1, 2))

}
