% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{fun_zoom}
\alias{fun_zoom}
\title{Disable zoom}
\usage{
fun_zoom(p, zoom = TRUE)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{zoom}{Set to \code{TRUE} to disable (default).}
}
\description{
Disable zoom.
}
\examples{
funplot() \%>\%
  fun_add("tan(x)") \%>\%
  fun_zoom()

}
