% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-plot.R
\name{funplot}
\alias{funplot}
\title{Initialise}
\usage{
funplot(width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{id of container (optional).}
}
\description{
Initialise a funplot.
}
\examples{
funplot() \%>\%
  fun_add("sin(x)")

}
