% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDfWithParams.R
\name{makeDfWithParams}
\alias{makeDfWithParams}
\title{makeDfWithParams}
\usage{
makeDfWithParams(fun, params, x_points)
}
\arguments{
\item{fun}{要绘制的函数}

\item{params}{函数参数df，类似params = data.frame(alpha = 1:10, beta = 1:10)}

\item{x_points}{x轴上的取样点}
}
\value{
绘图用的df = data.frame(x, y, params:str)
}
\description{
根据参数，为ggplot生成某个函数的绘图用的df = data.frame(x, y, params:str)
}
