% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotFunctions}
\alias{plotFunctions}
\title{plotFunctions}
\usage{
plotFunctions(fun, params = NULL, n = 101, x = c(1, 5), size = 0.5,
  end_spaces_ratio = 0)
}
\arguments{
\item{fun}{要绘制的函数}

\item{params}{函数的参数}

\item{n}{采样点}

\item{x}{x的区间}

\item{size}{线宽}

\item{end_spaces_ratio}{最右侧留空百分比（占x轴的总长度）}
}
\value{
ggplot2对象
}
\description{
绘制带参数的函数（组）
}
