% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fun_points}
\alias{fun_points}
\alias{fun_lines}
\title{Add points}
\usage{
fun_points(p, data, ...)

fun_lines(p, data, ...)
}
\arguments{
\item{p}{Plot as initialised by \code{\link{funplot}}.}

\item{data}{\code{data.frame} or \code{matrix} where the first
column is \eqn{x} and the sencond \eqn{y}.}

\item{...}{Any other parameter.}
}
\description{
Add points to plot.
}
\examples{
points <- data.frame(x = rnorm(10), y = rnorm(10))
lines <- data.frame(x = seq(-5, 4), y = rnorm(10))

funplot() \%>\%
  fun_points(points) \%>\%
  fun_lines(lines)

}
